changed qint16 to qreal so comma values are not cut off

+++ dashboard.cpp
@@ -59,7 +59,7 @@ DashBoard::DashBoard(QObject *parent)
 {
 }
 // Advanced Info FD3S
-void DashBoard::setRevs(const qint16 &revs)
+void DashBoard::setRevs(const qreal &revs)
 {
     if (m_revs == revs)
         return;
@@ -67,7 +67,7 @@ void DashBoard::setRevs(const qint16 &revs)
     emit revsChanged(revs);
 }
 
-void DashBoard::setIntakepress(const qint16 &Intakepress)
+void DashBoard::setIntakepress(const qreal &Intakepress)
 {
     if (m_Intakepress == Intakepress)
         return;
@@ -75,7 +75,7 @@ void DashBoard::setIntakepress(const qint16 &Intakepress)
     emit IntakepressChanged(Intakepress);
 }
 
-void DashBoard::setPressureV(const qint16 &PressureV)
+void DashBoard::setPressureV(const qreal &PressureV)
 {
     if (m_PressureV == PressureV)
         return;
@@ -83,7 +83,7 @@ void DashBoard::setPressureV(const qint16 &PressureV)
     emit PressureVChanged(PressureV);
 }
 
-void DashBoard::setThrottleV(const qint16 &ThrottleV)
+void DashBoard::setThrottleV(const qreal &ThrottleV)
 {
     if (m_ThrottleV == ThrottleV)
         return;
@@ -91,7 +91,7 @@ void DashBoard::setThrottleV(const qint16 &ThrottleV)
     emit ThrottleVChanged(ThrottleV);
 }
 
-void DashBoard::setPrimaryinp(const qint16 &Primaryinp)
+void DashBoard::setPrimaryinp(const qreal &Primaryinp)
 {
     if (m_Primaryinp == Primaryinp)
         return;
@@ -99,7 +99,7 @@ void DashBoard::setPrimaryinp(const qint16 &Primaryinp)
     emit PrimaryinpChanged(Primaryinp);
 }
 
-void DashBoard::setFuelc(const qint16 &Fuelc)
+void DashBoard::setFuelc(const qreal &Fuelc)
 {
     if (m_Fuelc == Fuelc)
         return;
@@ -107,7 +107,7 @@ void DashBoard::setFuelc(const qint16 &Fuelc)
     emit FuelcChanged(Fuelc);
 }
 
-void DashBoard::setLeadingign(const qint16 &Leadingign)
+void DashBoard::setLeadingign(const qreal &Leadingign)
 {
     if (m_Leadingign == Leadingign)
         return;
@@ -115,7 +115,7 @@ void DashBoard::setLeadingign(const qint16 &Leadingign)
     emit LeadingignChanged(Leadingign);
 }
 
-void DashBoard::setTrailingign(const qint16 &Trailingign)
+void DashBoard::setTrailingign(const qreal &Trailingign)
 {
     if (m_Trailingign == Trailingign)
         return;
@@ -123,7 +123,7 @@ void DashBoard::setTrailingign(const qint16 &Trailingign)
     emit TrailingignChanged(Trailingign);
 }
 
-void DashBoard::setFueltemp(const qint16 &Fueltemp)
+void DashBoard::setFueltemp(const qreal &Fueltemp)
 {
     if (m_Fueltemp == Fueltemp)
         return;
@@ -131,7 +131,7 @@ void DashBoard::setFueltemp(const qint16 &Fueltemp)
     emit FueltempChanged(Fueltemp);
 }
 
-void DashBoard::setMoilp(const qint16 &Moilp)
+void DashBoard::setMoilp(const qreal &Moilp)
 {
     if (m_Moilp == Moilp)
         return;
@@ -139,7 +139,7 @@ void DashBoard::setMoilp(const qint16 &Moilp)
     emit MoilpChanged(Moilp);
 }
 
-void DashBoard::setBoosttp(const qint16 &Boosttp)
+void DashBoard::setBoosttp(const qreal &Boosttp)
 {
     if (m_Boosttp == Boosttp)
         return;
@@ -147,7 +147,7 @@ void DashBoard::setBoosttp(const qint16 &Boosttp)
     emit BoosttpChanged(Boosttp);
 }
 
-void DashBoard::setBoostwg(const qint16 &Boostwg)
+void DashBoard::setBoostwg(const qreal &Boostwg)
 {
     if (m_Boostwg == Boostwg)
         return;
@@ -155,7 +155,7 @@ void DashBoard::setBoostwg(const qint16 &Boostwg)
     emit BoostwgChanged(Boostwg);
 }
 
-void DashBoard::setWatertemp(const qint16 &Watertemp)
+void DashBoard::setWatertemp(const qreal &Watertemp)
 {
     if (m_Watertemp == Watertemp)
         return;
@@ -163,7 +163,7 @@ void DashBoard::setWatertemp(const qint16 &Watertemp)
     emit WatertempChanged(Watertemp);
 }
 
-void DashBoard::setIntaketemp(const qint16 &Intaketemp)
+void DashBoard::setIntaketemp(const qreal &Intaketemp)
 {
     if (m_Intaketemp == Intaketemp)
         return;
@@ -171,7 +171,7 @@ void DashBoard::setIntaketemp(const qint16 &Intaketemp)
     emit IntaketempChanged(Intaketemp);
 }
 
-void DashBoard::setKnock(const qint16 &Knock)
+void DashBoard::setKnock(const qreal &Knock)
 {
     if (m_Knock == Knock)
         return;
@@ -179,7 +179,7 @@ void DashBoard::setKnock(const qint16 &Knock)
     emit KnockChanged(Knock);
 }
 
-void DashBoard::setBatteryV(const qint16 &BatteryV)
+void DashBoard::setBatteryV(const qreal &BatteryV)
 {
     if (m_BatteryV == BatteryV)
         return;
@@ -187,7 +187,7 @@ void DashBoard::setBatteryV(const qint16 &BatteryV)
     emit BatteryVChanged(BatteryV);
 }
 
-void DashBoard::setSpeed(const qint16 &speed)
+void DashBoard::setSpeed(const qreal &speed)
 {
     if (m_speed == speed)
         return;
@@ -195,7 +195,7 @@ void DashBoard::setSpeed(const qint16 &speed)
     emit speedChanged(speed);
 }
 
-void DashBoard::setIscvduty(const qint16 &Iscvduty)
+void DashBoard::setIscvduty(const qreal &Iscvduty)
 {
     if (m_Iscvduty == Iscvduty)
         return;
@@ -203,14 +203,14 @@ void DashBoard::setIscvduty(const qint16 &Iscvduty)
     emit IscvdutyChanged(Iscvduty);
 }
 
-void DashBoard::setO2volt(const qint16 &O2volt)
+void DashBoard::setO2volt(const qreal &O2volt)
 {
     if (m_O2volt == O2volt)
         return;
     m_O2volt = O2volt;
     emit O2voltChanged(O2volt);
 }
-void DashBoard::setna1(const qint16 &na1)
+void DashBoard::setna1(const qreal &na1)
 {
     if (m_na1 == na1)
         return;
@@ -218,7 +218,7 @@ void DashBoard::setna1(const qint16 &na1)
     emit na1Changed(na1);
 }
 
-void DashBoard::setSecinjpulse(const qint16 &Secinjpulse)
+void DashBoard::setSecinjpulse(const qreal &Secinjpulse)
 {
     if (m_Secinjpulse == Secinjpulse)
         return;
@@ -226,7 +226,7 @@ void DashBoard::setSecinjpulse(const qint16 &Secinjpulse)
     emit SecinjpulseChanged(Secinjpulse);
 }
 
-void DashBoard::setna2(const qint16 &na2)
+void DashBoard::setna2(const qreal &na2)
 {
     if (m_na2 == na2)
         return;
@@ -236,7 +236,7 @@ void DashBoard::setna2(const qint16 &na2)
 
 //Sensor info
 
-void DashBoard::setpim(const qint16 &pim)
+void DashBoard::setpim(const qreal &pim)
 {
     if (m_pim == pim)
         return;
@@ -244,7 +244,7 @@ void DashBoard::setpim(const qint16 &pim)
     emit pimChanged(pim);
 }
 
-void DashBoard::setvta1(const qint16 &vta1)
+void DashBoard::setvta1(const qreal &vta1)
 {
     if (m_vta1 == vta1)
         return;
@@ -252,7 +252,7 @@ void DashBoard::setvta1(const qint16 &vta1)
     emit vta1Changed(vta1);
 }
 
-void DashBoard::setvta2(const qint16 &vta2)
+void DashBoard::setvta2(const qreal &vta2)
 {
     if (m_vta2 == vta2)
         return;
@@ -260,7 +260,7 @@ void DashBoard::setvta2(const qint16 &vta2)
     emit vta2Changed(vta2);
 }
 
-void DashBoard::setvmop(const qint16 &vmop)
+void DashBoard::setvmop(const qreal &vmop)
 {
     if (m_vmop == vmop)
         return;
@@ -268,7 +268,7 @@ void DashBoard::setvmop(const qint16 &vmop)
     emit vmopChanged(vmop);
 }
 
-void DashBoard::setwtrt(const qint16 &wtrt)
+void DashBoard::setwtrt(const qreal &wtrt)
 {
     if (m_wtrt == wtrt)
         return;
@@ -276,7 +276,7 @@ void DashBoard::setwtrt(const qint16 &wtrt)
     emit wtrtChanged(wtrt);
 }
 
-void DashBoard::setairt(const qint16 &airt)
+void DashBoard::setairt(const qreal &airt)
 {
     if (m_airt == airt)
         return;
@@ -284,7 +284,7 @@ void DashBoard::setairt(const qint16 &airt)
     emit airtChanged(airt);
 }
 
-void DashBoard::setfuelt(const qint16 &fuelt)
+void DashBoard::setfuelt(const qreal &fuelt)
 {
     if (m_fuelt == fuelt)
         return;
@@ -292,7 +292,7 @@ void DashBoard::setfuelt(const qint16 &fuelt)
     emit fueltChanged(fuelt);
 }
 
-void DashBoard::setO2S(const qint16 &O2S)
+void DashBoard::setO2S(const qreal &O2S)
 {
     if (m_O2S == O2S)
         return;
@@ -304,7 +304,7 @@ void DashBoard::setO2S(const qint16 &O2S)
 
 //Flags
 
-void DashBoard::setFlag1(const qint16 &Flag1)
+void DashBoard::setFlag1(const qreal &Flag1)
 {
     if (m_Flag1 == Flag1)
         return;
@@ -312,7 +312,7 @@ void DashBoard::setFlag1(const qint16 &Flag1)
     emit Flag1Changed(Flag1);
 }
 
-void DashBoard::setFlag2(const qint16 &Flag2)
+void DashBoard::setFlag2(const qreal &Flag2)
 {
     if (m_Flag2 == Flag2)
         return;
@@ -320,7 +320,7 @@ void DashBoard::setFlag2(const qint16 &Flag2)
     emit Flag2Changed(Flag2);
 }
 
-void DashBoard::setFlag3(const qint16 &Flag3)
+void DashBoard::setFlag3(const qreal &Flag3)
 {
     if (m_Flag3 == Flag3)
         return;
@@ -328,7 +328,7 @@ void DashBoard::setFlag3(const qint16 &Flag3)
     emit Flag3Changed(Flag3);
 }
 
-void DashBoard::setFlag4(const qint16 &Flag4)
+void DashBoard::setFlag4(const qreal &Flag4)
 {
     if (m_Flag4 == Flag4)
         return;
@@ -336,7 +336,7 @@ void DashBoard::setFlag4(const qint16 &Flag4)
     emit Flag4Changed(Flag4);
 }
 
-void DashBoard::setFlag5(const qint16 &Flag5)
+void DashBoard::setFlag5(const qreal &Flag5)
 {
     if (m_Flag5 == Flag5)
         return;
@@ -344,7 +344,7 @@ void DashBoard::setFlag5(const qint16 &Flag5)
     emit Flag5Changed(Flag5);
 }
 
-void DashBoard::setFlag6(const qint16 &Flag6)
+void DashBoard::setFlag6(const qreal &Flag6)
 {
     if (m_Flag6 == Flag6)
         return;
@@ -352,7 +352,7 @@ void DashBoard::setFlag6(const qint16 &Flag6)
     emit Flag6Changed(Flag6);
 }
 
-void DashBoard::setFlag7(const qint16 &Flag7)
+void DashBoard::setFlag7(const qreal &Flag7)
 {
     if (m_Flag7 == Flag7)
         return;
@@ -360,7 +360,7 @@ void DashBoard::setFlag7(const qint16 &Flag7)
     emit Flag7Changed(Flag7);
 }
 
-void DashBoard::setFlag8(const qint16 &Flag8)
+void DashBoard::setFlag8(const qreal &Flag8)
 {
     if (m_Flag8 == Flag8)
         return;
@@ -368,7 +368,7 @@ void DashBoard::setFlag8(const qint16 &Flag8)
     emit Flag8Changed(Flag8);
 }
 
-void DashBoard::setFlag9(const qint16 &Flag9)
+void DashBoard::setFlag9(const qreal &Flag9)
 {
     if (m_Flag9 == Flag9)
         return;
@@ -376,7 +376,7 @@ void DashBoard::setFlag9(const qint16 &Flag9)
     emit Flag9Changed(Flag9);
 }
 
-void DashBoard::setFlag10(const qint16 &Flag10)
+void DashBoard::setFlag10(const qreal &Flag10)
 {
     if (m_Flag10 == Flag10)
         return;
@@ -384,7 +384,7 @@ void DashBoard::setFlag10(const qint16 &Flag10)
     emit Flag10Changed(Flag10);
 }
 
-void DashBoard::setFlag11(const qint16 &Flag11)
+void DashBoard::setFlag11(const qreal &Flag11)
 {
     if (m_Flag11 == Flag11)
         return;
@@ -392,7 +392,7 @@ void DashBoard::setFlag11(const qint16 &Flag11)
     emit Flag11Changed(Flag11);
 }
 
-void DashBoard::setFlag12(const qint16 &Flag12)
+void DashBoard::setFlag12(const qreal &Flag12)
 {
     if (m_Flag12 == Flag12)
         return;
@@ -400,7 +400,7 @@ void DashBoard::setFlag12(const qint16 &Flag12)
     emit Flag12Changed(Flag12);
 }
 
-void DashBoard::setFlag13(const qint16 &Flag13)
+void DashBoard::setFlag13(const qreal &Flag13)
 {
     if (m_Flag13 == Flag13)
         return;
@@ -408,7 +408,7 @@ void DashBoard::setFlag13(const qint16 &Flag13)
     emit Flag13Changed(Flag13);
 }
 
-void DashBoard::setFlag14(const qint16 &Flag14)
+void DashBoard::setFlag14(const qreal &Flag14)
 {
     if (m_Flag14 == Flag14)
         return;
@@ -416,7 +416,7 @@ void DashBoard::setFlag14(const qint16 &Flag14)
     emit Flag14Changed(Flag14);
 }
 
-void DashBoard::setFlag15(const qint16 &Flag15)
+void DashBoard::setFlag15(const qreal &Flag15)
 {
     if (m_Flag15 == Flag15)
         return;
@@ -424,7 +424,7 @@ void DashBoard::setFlag15(const qint16 &Flag15)
     emit Flag15Changed(Flag15);
 }
 
-void DashBoard::setFlag16(const qint16 &Flag16)
+void DashBoard::setFlag16(const qreal &Flag16)
 {
     if (m_Flag16 == Flag16)
         return;
@@ -437,59 +437,59 @@ void DashBoard::setFlag16(const qint16 &Flag16)
 
 
 // Advanced Info FD3S
-qint16 DashBoard::revs() const { return m_revs; }
-qint16 DashBoard::Intakepress() const { return m_Intakepress; }
-qint16 DashBoard::PressureV() const { return m_PressureV; }
-qint16 DashBoard::ThrottleV() const { return m_ThrottleV; }
-qint16 DashBoard::Primaryinp() const { return m_Primaryinp; }
-qint16 DashBoard::Fuelc() const { return m_Fuelc; }
-qint16 DashBoard::Leadingign() const { return m_Leadingign; }
-qint16 DashBoard::Trailingign() const { return m_Trailingign; }
-qint16 DashBoard::Fueltemp() const { return m_Fueltemp; }
-qint16 DashBoard::Moilp() const { return m_Moilp; }
-qint16 DashBoard::Boosttp() const { return m_Boosttp; }
-qint16 DashBoard::Boostwg() const { return m_Boostwg; }
-qint16 DashBoard::Watertemp() const { return m_Watertemp; }
-qint16 DashBoard::Intaketemp() const { return m_Intaketemp; }
-qint16 DashBoard::Knock() const { return m_Knock; }
-qint16 DashBoard::BatteryV() const { return m_BatteryV; }
-qint16 DashBoard::speed() const { return m_speed; }
-qint16 DashBoard::Iscvduty() const { return m_Iscvduty; }
-qint16 DashBoard::O2volt() const { return m_O2volt; }
-qint16 DashBoard::na1() const { return m_na1; }
-qint16 DashBoard::Secinjpulse() const { return m_Secinjpulse; }
-qint16 DashBoard::na2() const { return m_na2; }
+qreal DashBoard::revs() const { return m_revs; }
+qreal DashBoard::Intakepress() const { return m_Intakepress; }
+qreal DashBoard::PressureV() const { return m_PressureV; }
+qreal DashBoard::ThrottleV() const { return m_ThrottleV; }
+qreal DashBoard::Primaryinp() const { return m_Primaryinp; }
+qreal DashBoard::Fuelc() const { return m_Fuelc; }
+qreal DashBoard::Leadingign() const { return m_Leadingign; }
+qreal DashBoard::Trailingign() const { return m_Trailingign; }
+qreal DashBoard::Fueltemp() const { return m_Fueltemp; }
+qreal DashBoard::Moilp() const { return m_Moilp; }
+qreal DashBoard::Boosttp() const { return m_Boosttp; }
+qreal DashBoard::Boostwg() const { return m_Boostwg; }
+qreal DashBoard::Watertemp() const { return m_Watertemp; }
+qreal DashBoard::Intaketemp() const { return m_Intaketemp; }
+qreal DashBoard::Knock() const { return m_Knock; }
+qreal DashBoard::BatteryV() const { return m_BatteryV; }
+qreal DashBoard::speed() const { return m_speed; }
+qreal DashBoard::Iscvduty() const { return m_Iscvduty; }
+qreal DashBoard::O2volt() const { return m_O2volt; }
+qreal DashBoard::na1() const { return m_na1; }
+qreal DashBoard::Secinjpulse() const { return m_Secinjpulse; }
+qreal DashBoard::na2() const { return m_na2; }
 
 
 
 //Sensor info
 
-qint16 DashBoard::pim() const { return m_pim; }
-qint16 DashBoard::vta1() const { return m_vta1; }
-qint16 DashBoard::vta2() const { return m_vta2; }
-qint16 DashBoard::vmop() const { return m_vmop; }
-qint16 DashBoard::wtrt() const { return m_wtrt; }
-qint16 DashBoard::airt() const { return m_airt; }
-qint16 DashBoard::fuelt() const { return m_fuelt; }
-qint16 DashBoard::O2S() const { return m_O2S; }
+qreal DashBoard::pim() const { return m_pim; }
+qreal DashBoard::vta1() const { return m_vta1; }
+qreal DashBoard::vta2() const { return m_vta2; }
+qreal DashBoard::vmop() const { return m_vmop; }
+qreal DashBoard::wtrt() const { return m_wtrt; }
+qreal DashBoard::airt() const { return m_airt; }
+qreal DashBoard::fuelt() const { return m_fuelt; }
+qreal DashBoard::O2S() const { return m_O2S; }
 
 //Flags
 
-qint16 DashBoard::Flag1() const { return m_Flag1; }
-qint16 DashBoard::Flag2() const { return m_Flag2; }
-qint16 DashBoard::Flag3() const { return m_Flag3; }
-qint16 DashBoard::Flag4() const { return m_Flag4; }
-qint16 DashBoard::Flag5() const { return m_Flag5; }
-qint16 DashBoard::Flag6() const { return m_Flag6; }
-qint16 DashBoard::Flag7() const { return m_Flag7; }
-qint16 DashBoard::Flag8() const { return m_Flag8; }
-qint16 DashBoard::Flag9() const { return m_Flag9; }
-qint16 DashBoard::Flag10() const { return m_Flag10; }
-qint16 DashBoard::Flag11() const { return m_Flag11; }
-qint16 DashBoard::Flag12() const { return m_Flag12; }
-qint16 DashBoard::Flag13() const { return m_Flag13; }
-qint16 DashBoard::Flag14() const { return m_Flag14; }
-qint16 DashBoard::Flag15() const { return m_Flag15; }
-qint16 DashBoard::Flag16() const { return m_Flag16; }
+qreal DashBoard::Flag1() const { return m_Flag1; }
+qreal DashBoard::Flag2() const { return m_Flag2; }
+qreal DashBoard::Flag3() const { return m_Flag3; }
+qreal DashBoard::Flag4() const { return m_Flag4; }
+qreal DashBoard::Flag5() const { return m_Flag5; }
+qreal DashBoard::Flag6() const { return m_Flag6; }
+qreal DashBoard::Flag7() const { return m_Flag7; }
+qreal DashBoard::Flag8() const { return m_Flag8; }
+qreal DashBoard::Flag9() const { return m_Flag9; }
+qreal DashBoard::Flag10() const { return m_Flag10; }
+qreal DashBoard::Flag11() const { return m_Flag11; }
+qreal DashBoard::Flag12() const { return m_Flag12; }
+qreal DashBoard::Flag13() const { return m_Flag13; }
+qreal DashBoard::Flag14() const { return m_Flag14; }
+qreal DashBoard::Flag15() const { return m_Flag15; }
+qreal DashBoard::Flag16() const { return m_Flag16; }
